+++ src/main/java/artifacts/client/render/trinket/renderer/BeltCurioRenderer.java
@@ -2,6 +2,7 @@ package artifacts.client.render.trinket.renderer;
 
 import artifacts.Artifacts;
 import artifacts.client.render.trinket.model.BeltModel;
+import artifacts.common.trinkets.TrinketsHelper;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import dev.emi.trinkets.api.SlotReference;
@@ -38,7 +39,11 @@ public class BeltCurioRenderer implements TrinketRenderer {
     }
 
     @Override
-    public void render(ItemStack stack, SlotReference slotReference, EntityModel<? extends LivingEntity> contextModel, PoseStack poseStack, MultiBufferSource multiBufferSource, int light, LivingEntity entity, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch) {
+    public final void render(ItemStack stack, SlotReference slotReference, EntityModel<? extends LivingEntity> contextModel, PoseStack poseStack, MultiBufferSource multiBufferSource, int light, LivingEntity entity, float limbSwing, float limbSwingAmount, float partialTicks, float ageInTicks, float netHeadYaw, float headPitch) {
+        if (!TrinketsHelper.areCosmetisEnabled(stack)) {
+            return;
+        }
+
         BeltModel model = getModel();
 
         model.setupAnim(entity, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch);